requirements.txt
aiosqlite==0.19.0
annotated-types==0.6.0
anyio==3.7.1
apturl==0.5.2
asgiref==3.5.0
async-timeout==4.0.3
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==41.0.4
cupshelpers==1.0
databases==0.8.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.4.2
duplicity==0.8.21
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.2
fasteners==0.14.1
future==0.18.2
greenlet==3.0.0
h11==0.13.0
hiredis==2.2.3
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==9.1.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
pydantic==2.0.3
pydantic-settings==2.0.3
pydantic_core==2.3.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-ulid==1.1.0
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
redis==4.6.0
redis-om==0.2.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
systemd-python==234
types-pyOpenSSL==23.2.0.2
types-redis==4.6.0.7
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.15.0
wadllib==1.3.6
wsproto==1.0.0
xdg==5
xkit==0.0.0
youtube-dl==2021.12.17
zipp==1.0.0
